Doctor Help- Project completed in May 2013
Invigilation software(project completed for database class in May 2013) assigns invigilation duties to the teachers of a department for all the exams conducted during a semester. It takes into account the seniority of a teacher, their time schedules, examination time tables and all such various factors while assigning the duties. It provides a dual interface – one for the faculty and the other for the administrator. The software has been developed using Microsoft Visual Studio 2010 at the front end and Oracle at the back end.